Text Updates
Updated various text things since I am behind on the times.
Changed the display name to be the new Dragon Ancestry and Lineages 9000 name.
Changed all instances of 'gender' to 'mating type' because dragons are fungi.
Added the new genes that happened since the last time I updated the genes.
Attempted to add the missing dragon image.  We'll see how that does.